Add Entity class to trim Unit class of generics.
